Add basic cufft example, use double instead of float

+++ hpp/typedefs.hpp
@@ -7,7 +7,7 @@ namespace cuda_tutorial {
 
 using size = std::size_t;
 using ssize = std::ptrdiff_t;
-using float_pt = float;
+using float_pt = double;
 
 struct work_division {
     dim3 blocks;
